Adding a WIP policy server and simple system clock component for easy testing.

+++ component-heating/src/main/java/com/furnaghan/home/component/heating/HeatingType.java
@@ -10,5 +10,5 @@ public interface HeatingType extends MeterType {
 
     void turnOff();
 
-    void setTemperature(final String zone, final float temperature);
+    void setTemperature(final String zone, final double temperature);
 }
